Revert "SOLR-13350: Merging master and updating the PR"
This reverts commit afd6f5a139cd8b5b16212dc9cc564656e64f09a3.

+++ solr/contrib/analytics/src/test/org/apache/solr/analytics/legacy/LegacyNoFacetCloudTest.java
@@ -131,22 +131,22 @@ public class LegacyNoFacetCloudTest extends LegacyAbstractAnalyticsCloudTest {
 
     //Int
     Double intResult = getValue(response, "sr", "int_id");
-    Double intTest = calculateNumberStat(intTestStart, "sum");
+    Double intTest = (Double)calculateNumberStat(intTestStart, "sum");
     assertEquals(responseStr, intResult,intTest);
 
     //Long
     Double longResult = getValue(response, "sr", "long_ld");
-    Double longTest = calculateNumberStat(longTestStart, "sum");
+    Double longTest = (Double)calculateNumberStat(longTestStart, "sum");
     assertEquals(responseStr, longResult,longTest);
 
     //Float
     Double floatResult = getValue(response, "sr", "float_fd");
-    Double floatTest = calculateNumberStat(floatTestStart, "sum");
+    Double floatTest = (Double)calculateNumberStat(floatTestStart, "sum");
     assertEquals(responseStr, floatResult,floatTest);
 
     //Double
     Double doubleResult = getValue(response, "sr", "double_dd");
-    Double doubleTest = calculateNumberStat(doubleTestStart, "sum");
+    Double doubleTest = (Double) calculateNumberStat(doubleTestStart, "sum");
     assertEquals(responseStr, doubleResult,doubleTest);
   }
 
@@ -163,22 +163,22 @@ public class LegacyNoFacetCloudTest extends LegacyAbstractAnalyticsCloudTest {
 
     //Int
     Double intResult = getValue(response, "mr", "int_id");
-    Double intTest = calculateNumberStat(intTestStart, "mean");
+    Double intTest = (Double)calculateNumberStat(intTestStart, "mean");
     assertEquals(responseStr, intResult,intTest);
 
     //Long
     Double longResult = getValue(response, "mr", "long_ld");
-    Double longTest = calculateNumberStat(longTestStart, "mean");
+    Double longTest = (Double)calculateNumberStat(longTestStart, "mean");
     assertEquals(responseStr, longResult,longTest);
 
     //Float
     Double floatResult = getValue(response, "mr", "float_fd");
-    Double floatTest = calculateNumberStat(floatTestStart, "mean");
+    Double floatTest = (Double)calculateNumberStat(floatTestStart, "mean");
     assertEquals(responseStr, floatResult,floatTest);
 
     //Double
     Double doubleResult = getValue(response, "mr", "double_dd");
-    Double doubleTest = calculateNumberStat(doubleTestStart, "mean");
+    Double doubleTest = (Double)calculateNumberStat(doubleTestStart, "mean");
     assertEquals(responseStr, doubleResult,doubleTest);
   }
 
@@ -195,23 +195,23 @@ public class LegacyNoFacetCloudTest extends LegacyAbstractAnalyticsCloudTest {
 
     //Int
     Double intResult = getValue(response, "str", "int_id");
-    Double intTest = calculateNumberStat(intTestStart, "stddev");
+    Double intTest = (Double)calculateNumberStat(intTestStart, "stddev");
     assertEquals(responseStr, intResult, intTest, 0.00000000001);
 
     //Long
     Double longResult = getValue(response, "str", "long_ld");
-    Double longTest = calculateNumberStat(longTestStart, "stddev");
+    Double longTest = (Double)calculateNumberStat(longTestStart, "stddev");
     assertEquals(responseStr, longResult, longTest, 0.00000000001);
 
     //Float
     Double floatResult = getValue(response, "str", "float_fd");
-    Double floatTest = calculateNumberStat(floatTestStart, "stddev");
+    Double floatTest = (Double)calculateNumberStat(floatTestStart, "stddev");
     assertEquals(responseStr, floatResult, floatTest, 0.00000000001);
 
 
     //Double
     Double doubleResult = getValue(response, "str", "double_dd");
-    Double doubleTest = calculateNumberStat(doubleTestStart, "stddev");
+    Double doubleTest = (Double)calculateNumberStat(doubleTestStart, "stddev");
     assertEquals(responseStr, doubleResult, doubleTest, 0.00000000001);
   }
 
@@ -229,22 +229,22 @@ public class LegacyNoFacetCloudTest extends LegacyAbstractAnalyticsCloudTest {
 
     //Int
     Double intResult = getValue(response, "medr", "int_id");
-    Double intTest = calculateNumberStat(intTestStart, "median");
+    Double intTest = (Double)calculateNumberStat(intTestStart, "median");
     assertEquals(responseStr, intResult,intTest);
 
     //Long
     Double longResult = getValue(response, "medr", "long_ld");
-    Double longTest = calculateNumberStat(longTestStart, "median");
+    Double longTest = (Double)calculateNumberStat(longTestStart, "median");
     assertEquals(responseStr, longResult,longTest);
 
     //Float
     Double floatResult = getValue(response, "medr", "float_fd");
-    Double floatTest = calculateNumberStat(floatTestStart, "median");
+    Double floatTest = (Double)calculateNumberStat(floatTestStart, "median");
     assertEquals(responseStr, floatResult,floatTest);
 
     //Double
     Double doubleResult = getValue(response, "medr", "double_dd");
-    Double doubleTest = calculateNumberStat(doubleTestStart, "median");
+    Double doubleTest = (Double)calculateNumberStat(doubleTestStart, "median");
     assertEquals(responseStr, doubleResult,doubleTest);
 
     // TODO: Add test for date median
